+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.2] - 2023-12-17
+
 ### Added
  - `ArraySubset::{extract_elements/extract_elements_unchecked}` and `ArrayExtractElementsError`
 
@@ -308,7 +310,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - Initial public release
 
-[unreleased]: https://github.com/LDeakin/zarrs/compare/v0.7.1...HEAD
+[unreleased]: https://github.com/LDeakin/zarrs/compare/v0.7.2...HEAD
+[0.7.2]: https://github.com/LDeakin/zarrs/releases/tag/v0.7.2
 [0.7.1]: https://github.com/LDeakin/zarrs/releases/tag/v0.7.1
 [0.7.0]: https://github.com/LDeakin/zarrs/releases/tag/v0.7.0
 [0.6.0]: https://github.com/LDeakin/zarrs/releases/tag/v0.6.0